+++ RELEASE_NOTES.md
@@ -4,6 +4,8 @@ This file contains one line summaries of commits that are worthy of mentioning i
 A new header is inserted each time a *tag* is created.
 
 ## main branch
+ 
+- engine: Add optional memory configuration parameters to Engine initialization
 
 ## v1.25.2